+++ CotEditor/Sources/PreferencesTabViewController.swift
@@ -110,25 +110,13 @@ final class PreferencesTabViewController: NSTabViewController {
         frame.origin = window.frame.origin
         frame.origin.y += window.frame.height - frame.height
         
-        if #available(macOS 10.13, *) {
-            self.view.isHidden = true
-        } else {
-            // use `alphaValue` instead of `isHidden`
-            // -> Because updating `isHidden` in completionHandler under macOS 10.12 causes
-            //    either crash or skip initial update of the theme editor (although OK on 10.13 and above).
-            self.view.alphaValue = 0
-        }
+        self.view.isHidden = true
         NSAnimationContext.runAnimationGroup({ context in
             context.allowsImplicitAnimation = animated
-            
             window.setFrame(frame, display: false)
             
         }, completionHandler: { [weak self] in
-            if #available(macOS 10.13, *) {
-                self?.view.isHidden = false
-            } else {
-                self?.view.alphaValue = 1
-            }
+            self?.view.isHidden = false
             window.title = tabViewItem.label
         })
     }