+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Change Log
 
 - Fix an issue where the final number of replacement in the progress dialog was occasionally underrated.
 - Fix an issue where the curusor could not be set at the end of the document via AppleScript or `cot` command.
+- Fix the preferences pane switching animation under macOS 10.13 and earlier.
 - [trivial] Fix a potential issue where last empty line number could be drawn at the first line position under specific condition.
 - [trivial] Fix an issue where editor scrolled occasionally unintendedly when an arrow key is pressed.